factor PeerStreamClient into PeerRPCClient
