+++ index.js
@@ -1,3 +1,2 @@
 module.exports.PeerRPCClient = require('./lib/PeerRPCClient')
 module.exports.PeerRPCServer = require('./lib/PeerRPCServer')
-module.exports.PeerRPCStreamClient = require('./lib/PeerRPCStreamClient')
