+++ lib/TransportRPCClient.js
@@ -32,12 +32,11 @@ class TransportRPCClient extends Base.TransportRPCClient {
     return request
   }
 
-  getOpts (postData, opts, secure) {
+  getOpts (opts, secure) {
     const u = secure
       ? ('https://' + this.conf.dest) : ('http://' + this.conf.dest)
 
     const def = {
-      body: postData,
       url: u,
       path: '/',
       method: 'POST'
@@ -58,8 +57,9 @@ class TransportRPCClient extends Base.TransportRPCClient {
     const postData = this.format([req.rid, req.key, req.payload])
 
     this.post({
-      timeout: req.opts.timeout
-    }, postData, (err, body) => {
+      timeout: req.opts.timeout,
+      body: postData
+    }, (err, body) => {
       if (err) {
         this.handleReply(req.rid, new Error(`ERR_REQUEST_GENERIC: ${err.message}`))
         return
@@ -77,9 +77,9 @@ class TransportRPCClient extends Base.TransportRPCClient {
     })
   }
 
-  post (_opts, postData, _cb) {
+  post (_opts, _cb) {
     const socket = this.socket
-    const opts = this.getOpts(postData, _opts, this.conf.secure)
+    const opts = this.getOpts(_opts, this.conf.secure)
 
     let isExecuted = false
 
@@ -100,6 +100,40 @@ class TransportRPCClient extends Base.TransportRPCClient {
     req.on('error', (err) => {
       cb(err)
     })
+  }
+
+  requestStream (key, opts, cb) {
+    return this._requestStream(key, opts, cb)
+  }
+
+  _requestStream (key, opts, cb) {
+    const req = this.newRequest(key, null, opts, cb)
+    this.addRequest(req)
+    return this.sendRequestStream(req)
+  }
+
+  sendRequestStream (req) {
+    const addHeaders = {}
+    const _h = req.opts.headers || {}
+    Object.keys(_h).forEach((k) => {
+      if (typeof _h[k] === 'string') {
+        addHeaders[k] = _h[k]
+        return
+      }
+      addHeaders[k] = JSON.stringify(_h[k])
+    })
+
+    const _opts = {
+      headers: {
+        _gr: this.format([req.rid, req.key]),
+        ...addHeaders
+      },
+      timeout: req.opts.timeout
+    }
+
+    const opts = this.getOpts(_opts, this.conf.secure)
+    const stream = this.socket.post(opts)
+    return stream
   }
 }
