+++ test/rpc-server-readstream-size.js
@@ -3,12 +3,12 @@
 const assert = require('assert')
 const fs = require('fs')
 const path = require('path')
-const { PeerRPCClient, PeerRPCServer, PeerRPCStreamClient } = require('../')
+const { PeerRPCClient, PeerRPCServer } = require('../')
 const Link = require('grenache-nodejs-link')
 const setupHooks = require('./before-after.js')
 
 const PORT = 1337
-let link, peer, peerSrvStr, serviceStr, pst, stop
+let link, peer, peerSrvStr, serviceStr, stop
 describe('RPC integration', () => {
   setupHooks((grapes) => {
     link = new Link({
@@ -18,9 +18,6 @@ describe('RPC integration', () => {
 
     peer = new PeerRPCClient(link, {})
     peer.init()
-
-    pst = new PeerRPCStreamClient(link, {})
-    pst.init()
 
     peerSrvStr = new PeerRPCServer(link, {
       timeout: 300000,
@@ -37,7 +34,6 @@ describe('RPC integration', () => {
 
     stop = () => {
       peer.stop()
-      pst.stop()
       link.stop()
       serviceStr.stop()
     }
@@ -69,7 +65,7 @@ describe('RPC integration', () => {
       bucket: 'BUCKET',
       contentType: 'application/pdf'
     }
-    const req = pst.stream('rpc_manual', {
+    const req = peer.stream('rpc_manual', {
       headers: {
         _a: 'uploadNewPublicStream',
         _ar: optsPdf,
